+++ Project/Src/Gui/InsightWindow/InsightWindow.cs
@@ -9,7 +9,6 @@ using System;
 using System.Collections.Generic;
 using System.Drawing;
 using System.Windows.Forms;
-
 using ICSharpCode.TextEditor.Gui.CompletionWindow;
 using ICSharpCode.TextEditor.Util;
 
@@ -61,7 +60,7 @@ namespace ICSharpCode.TextEditor.Gui.InsightWindow
         {
             // move the window under the caret (don't change the x position)
             TextLocation caretPos  = control.ActiveTextAreaControl.Caret.Position;
-            int y = (int)((1 + caretPos.Y) * control.ActiveTextAreaControl.TextArea.TextView.FontHeight)
+            int y = (1 + caretPos.Y) * control.ActiveTextAreaControl.TextArea.TextView.FontHeight
                 - control.ActiveTextAreaControl.TextArea.VirtualTop.Y - 1
                 + control.ActiveTextAreaControl.TextArea.TextView.DrawingPosition.Y;
